Fix Small Bugs For Public Release

Fixes two small bugs.

+++ sratool/sratool.h
@@ -163,7 +163,7 @@ void wlist_alist(void);
 void print_raw(const uint8_t *pkt, size_t pkt_len);
 bool print_dns_pkt(const uint8_t *data, size_t data_len, const char *str);
 void print_raw_ip(const uint8_t *data, size_t data_len, axa_p_ch_t ch);
-void print_stats(_axa_p_stats_rsp_t *stats, size_t len);
+void print_stats(_axa_p_stats_rsp_t *stats, uint32_t len);
 void print_kill(_axa_p_kill_t *kill, size_t len);
 
 /* server.c */
